+++ src/theme/createTheme.ts
@@ -1,6 +1,6 @@
-import { ITheme, IThemeColors, IThemeSpacing, IThemeMediaQueries, IThemeBaseColors, IThemeFont } from "./ITheme";
+import { ITheme, IThemeColors, IThemeSpacing, IThemeBaseColors, IThemeFont } from "./ITheme";
 import { IPalette } from "./IPalette";
-import { breakpoints } from "../layout/responsive/breakpoints";
+import { mediaQueries } from "../layout/responsive/mediaQueries";
 
 /*
  * We don't export a theme object directly to avoid importing it by accident without dependency injection
@@ -294,30 +294,6 @@ export const createTheme: (palette: IPalette) => ITheme = (palette) => {
         topbarHeight: 72
     };
 
-    let { xsMax, sMax, mMax, lMax, xlMax } = breakpoints;
-
-    let media: IThemeMediaQueries = {
-        xs: `screen and (max-width: ${xsMax - 1}px)`,
-
-        sAndBelow: `screen and (max-width: ${sMax - 1}px)`,
-        sExact: `screen and (min-width: ${xsMax}px) and (max-width: ${sMax - 1}px)`,
-        sAndAbove: `screen and (min-width: ${xsMax}px)`,
-
-        mAndBelow: `screen and (max-width: ${mMax - 1}px)`,
-        mExact: `screen and (min-width: ${sMax}px) and (max-width: ${mMax - 1}px)`,
-        mAndAbove: `screen and (min-width: ${sMax}px)`,
-
-        lAndBelow: `screen and (max-width: ${lMax - 1}px)`,
-        lExact: `screen and (min-width: ${mMax}px) and (max-width: ${lMax - 1}px)`,
-        lAndAbove: `screen and (min-width: ${mMax}px)`,
-
-        xlAndBelow: `screen and (max-width: ${xlMax - 1}px)`,
-        xlExact: `screen and (min-width: ${lMax}px) and (max-width: ${xlMax - 1}px)`,
-        xlAndAbove: `screen and (min-width: ${lMax}px)`,
-
-        xxl: `screen and (min-width: ${xlMax}px)`
-    };
-
     let font: IThemeFont = {
         weight: {
             thin: 100,
@@ -335,7 +311,7 @@ export const createTheme: (palette: IPalette) => ITheme = (palette) => {
     return {
         colors,
         spacing,
-        media,
+        media: mediaQueries,
         font
     };
 };