+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+- (layout/responsive/*): Add React hooks `useViewportSize` and `useMediaCheck` (with HOC alternative `WithMediaCheck`)
+- (ITheme): Predefined media queries are marked as non-customizable in the theme and also provided as a static export.
+
 # v2.0.0-dev.3
 
 - (PlusIcon): Fix SVG attribute casing